Last BCD ID 같을 경우 만들어서 보내기

+++ NotchingGradeInsp/CImageProcess.h
@@ -92,7 +92,7 @@ public:
 	int	nTabStartPosInFrame ; // Frame에서 Tab의 시작위치 (거리계산용)
 
 	//Grab Image 얻은 시점의 BCD ID
-	int m_GrabCallBCDId;
+	UINT64 m_GrabCallBCDId;
 
 	//Grab Call 호출 타임(ms)
 	UINT64 m_GrabCallTime;
